refactor: procject structure
