+++ reader.go
@@ -7,7 +7,10 @@ import (
 	"io"
 	"io/fs"
 	"log/slog"
+	"strconv"
+	"strings"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/aws/smithy-go"
@@ -15,49 +18,43 @@ import (
 
 // ObjectReader is an io.Reader implementation for an S3 Object
 type ObjectReader struct {
-	ctx         context.Context
-	cli         *s3.Client
-	bucket      string
-	key         string
-	rd          *io.PipeReader
-	logger      *slog.Logger
-	chunkSize   int
-	concurrency int
-	s3Opts      []func(*s3.Options)
+	ctx           context.Context
+	s3            DownloadAPIClient
+	rd            *io.PipeReader
+	logger        *slog.Logger
+	chunkSize     int64
+	concurrency   int
+	retries       int
+	clientOptions []func(*s3.Options)
+	input         *s3.GetObjectInput
 }
 
 // NewObjectReader returns a new ObjectReader to do io.Reader opperations on your s3 object
-func NewObjectReader(ctx context.Context, cli *s3.Client, bucketName, key string, opts ...ObjectReaderOption) (*ObjectReader, error) {
+func NewObjectReader(ctx context.Context, s3 DownloadAPIClient, input *s3.GetObjectInput, opts ...ObjectReaderOption) io.Reader {
 	rd := &ObjectReader{
 		ctx:         ctx,
-		cli:         cli,
-		bucket:      bucketName,
-		key:         key,
+		s3:          s3,
+		input:       input,
+		retries:     defaultRetries,
 		chunkSize:   DefaultChunkSize,
 		concurrency: defaultConcurrency,
 		logger:      noopLogger,
 	}
 
-	if err := ObjectReaderOptions(opts...)(rd); err != nil {
-		return nil, err
-	}
+	ObjectReaderOptions(opts...)(rd)
 
-	return rd, nil
+	return rd
 }
 
 // ObjectReaderOption is an option for the given read operation
-type ObjectReaderOption func(*ObjectReader) error
+type ObjectReaderOption func(*ObjectReader)
 
 // ObjectReaderOptions is a collection of ObjectReaderOption's
 func ObjectReaderOptions(opts ...ObjectReaderOption) ObjectReaderOption {
-	return func(r *ObjectReader) error {
+	return func(r *ObjectReader) {
 		for _, op := range opts {
-			if err := op(r); err != nil {
-				return err
-			}
+			op(r)
 		}
-
-		return nil
 	}
 }
 
@@ -86,10 +83,10 @@ func (r *ObjectReader) preRead() error {
 
 	r.rd = rd
 
-	res, err := r.cli.HeadObject(ctx, &s3.HeadObjectInput{
-		Bucket: &r.bucket,
-		Key:    &r.key,
-	}, r.s3Opts...)
+	input := *r.input
+	input.Range = aws.String("bytes=0-0")
+
+	res, err := r.s3.GetObject(ctx, &input, r.clientOptions...)
 	if err != nil {
 		var apiError smithy.APIError
 		if errors.As(err, &apiError) {
@@ -103,8 +100,31 @@ func (r *ObjectReader) preRead() error {
 			}
 		}
 	}
+	defer res.Body.Close()
 
-	contentLen := *res.ContentLength
+	var contentLen int64
+	if res.ContentRange == nil {
+		if l := aws.ToInt64(res.ContentLength); l > 0 {
+			contentLen = l
+		}
+	} else {
+		parts := strings.Split(*res.ContentRange, "/")
+
+		total := int64(-1)
+		var err error
+		// Checking for whether or not a numbered total exists
+		// If one does not exist, we will assume the total to be -1, undefined,
+		// and sequentially download each chunk until hitting a 416 error
+		totalStr := parts[len(parts)-1]
+		if totalStr != "*" {
+			total, err = strconv.ParseInt(totalStr, 10, 64)
+			if err != nil {
+				return err
+			}
+		}
+
+		contentLen = total
+	}
 
 	r.logger.Debug("pre read", slog.Int64("content-length", contentLen))
 	cl := newConcurrencyLock(r.concurrency)
@@ -175,13 +195,20 @@ func (r *ObjectReader) getObject(ctx context.Context, start, end int64) (*s3.Get
 	r.logger.DebugContext(ctx, "getting chunk", slog.Group("chunk", slog.Int64("start", start), slog.Int64("end", end)))
 
 	byteRange := fmt.Sprintf("bytes=%d-%d", start, end)
-	input := &s3.GetObjectInput{
-		Bucket: &r.bucket,
-		Key:    &r.key,
-		Range:  &byteRange,
+
+	input := *r.input
+	input.Range = &byteRange
+
+	var res *s3.GetObjectOutput
+	var err error
+	for range r.retries {
+		res, err = r.s3.GetObject(ctx, &input, r.clientOptions...)
+		if err == nil {
+			return res, err
+		}
 	}
 
-	return r.cli.GetObject(ctx, input, r.s3Opts...)
+	return res, err
 }
 
 /*
@@ -190,47 +217,45 @@ func (r *ObjectReader) getObject(ctx context.Context, start, end int64) (*s3.Get
 
 // WithReaderLogger sets the logger for this reader
 func WithReaderLogger(logger *slog.Logger) ObjectReaderOption {
-	return func(r *ObjectReader) error {
+	return func(r *ObjectReader) {
 		if logger != nil {
 			r.logger = logger
 		}
-
-		return nil
 	}
 }
 
 // WithReaderChunkSize sets the chunksize for this reader
-func WithReaderChunkSize(size uint) ObjectReaderOption {
-	return func(r *ObjectReader) error {
-		r.chunkSize = int(size)
-
-		return nil
+func WithReaderChunkSize(size int64) ObjectReaderOption {
+	return func(r *ObjectReader) {
+		r.chunkSize = size
 	}
 }
 
 // WithReaderConcurrency set the concurency amount for this reader
-func WithReaderConcurrency(i uint8) ObjectReaderOption {
-	return func(r *ObjectReader) error {
-		r.concurrency = int(i)
+func WithReaderConcurrency(i int) ObjectReaderOption {
+	return func(r *ObjectReader) {
+		if i < 1 {
+			i = 1
+		}
 
-		return nil
+		r.concurrency = i
 	}
 }
 
 // WithReaderRetries sets the retry count for this reader
-func WithReaderRetries(i uint8) ObjectReaderOption {
-	return func(r *ObjectReader) error {
-		r.s3Opts = append(r.s3Opts, withS3Retries(int(i)))
+func WithReaderRetries(i int) ObjectReaderOption {
+	return func(r *ObjectReader) {
+		if i < 1 {
+			i = 1
+		}
 
-		return nil
+		r.retries = i
 	}
 }
 
 // WithReaderS3Options adds s3 options to the reader opperations
 func WithReaderS3Options(opts ...func(*s3.Options)) ObjectReaderOption {
-	return func(r *ObjectReader) error {
-		r.s3Opts = append(r.s3Opts, opts...)
-
-		return nil
+	return func(r *ObjectReader) {
+		r.clientOptions = append(r.clientOptions, opts...)
 	}
 }
